chore: remove cli from playwright-core
