+++ packages/playwright-firefox/cli.js
@@ -14,4 +14,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-module.exports = require('playwright-core/cli');
+module.exports = require('playwright-core/lib/cli/cli');
